Refactor StructureInsn class
